+++ My project/Assets/Resources/Prefabs/Skills/SubChar/Kunai.prefab
@@ -12,7 +12,7 @@ GameObject:
   - component: {fileID: 6962129636200254812}
   m_Layer: 0
   m_Name: Kunai
-  m_TagString: Untagged
+  m_TagString: skill
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
