+++ My project/Assets/Resources/Prefabs/Skills/SubChar/dager.prefab
@@ -97,7 +97,7 @@ GameObject:
   - component: {fileID: -4434437990185676618}
   m_Layer: 0
   m_Name: dager
-  m_TagString: Untagged
+  m_TagString: SubAttack
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
