+++ My project/Assets/Scenes/Lee_Test.unity
@@ -251,6 +251,7 @@ MonoBehaviour:
   attackDmg: 0
   healing: 0
   defense: 0
+  this_name: 
   atkSpeed: 0
 --- !u!1001 &64197497
 PrefabInstance:
@@ -269,7 +270,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 733390301504948927, guid: 69ee51c36f135b549ada05bb9e55f926, type: 3}
       propertyPath: m_LocalPosition.x
-      value: 7.97
+      value: 12.01
       objectReference: {fileID: 0}
     - target: {fileID: 733390301504948927, guid: 69ee51c36f135b549ada05bb9e55f926, type: 3}
       propertyPath: m_LocalPosition.y
@@ -311,6 +312,14 @@ PrefabInstance:
       propertyPath: CreaterMonster
       value: 
       objectReference: {fileID: 64197499}
+    - target: {fileID: 7410440713986742281, guid: 69ee51c36f135b549ada05bb9e55f926, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7410440713986742281, guid: 69ee51c36f135b549ada05bb9e55f926, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 69ee51c36f135b549ada05bb9e55f926, type: 3}
 --- !u!114 &64197499 stripped
@@ -459,6 +468,22 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 214604645, guid: 129f1b5fe10ef424da0ff2abc5981855, type: 3}
+      propertyPath: m_TagString
+      value: skill
+      objectReference: {fileID: 0}
+    - target: {fileID: 538154444, guid: 129f1b5fe10ef424da0ff2abc5981855, type: 3}
+      propertyPath: m_TagString
+      value: skill
+      objectReference: {fileID: 0}
+    - target: {fileID: 1274880094, guid: 129f1b5fe10ef424da0ff2abc5981855, type: 3}
+      propertyPath: m_TagString
+      value: skill
+      objectReference: {fileID: 0}
+    - target: {fileID: 2109872032, guid: 129f1b5fe10ef424da0ff2abc5981855, type: 3}
+      propertyPath: m_TagString
+      value: skill
+      objectReference: {fileID: 0}
     - target: {fileID: 2314681825701584321, guid: 129f1b5fe10ef424da0ff2abc5981855, type: 3}
       propertyPath: m_RootOrder
       value: 1
@@ -507,6 +532,10 @@ PrefabInstance:
       propertyPath: m_Name
       value: Player
       objectReference: {fileID: 0}
+    - target: {fileID: 2314681825701584322, guid: 129f1b5fe10ef424da0ff2abc5981855, type: 3}
+      propertyPath: m_TagString
+      value: Player
+      objectReference: {fileID: 0}
     - target: {fileID: 2314681826405287298, guid: 129f1b5fe10ef424da0ff2abc5981855, type: 3}
       propertyPath: m_LocalPosition.x
       value: -0.56
@@ -528,6 +557,14 @@ PrefabInstance:
       propertyPath: m_Name
       value: Kunai
       objectReference: {fileID: 0}
+    - target: {fileID: 663929564521697501, guid: d0607e85f5e44774284628227ea994d4, type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 663929564521697501, guid: d0607e85f5e44774284628227ea994d4, type: 3}
+      propertyPath: m_TagString
+      value: skill
+      objectReference: {fileID: 0}
     - target: {fileID: 4691233938303236447, guid: d0607e85f5e44774284628227ea994d4, type: 3}
       propertyPath: m_LocalRotation.w
       value: 1
@@ -539,6 +576,10 @@ PrefabInstance:
     - target: {fileID: 4691233938303236447, guid: d0607e85f5e44774284628227ea994d4, type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8257185624544131300, guid: d0607e85f5e44774284628227ea994d4, type: 3}
+      propertyPath: m_IsActive
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 8755122303968071363, guid: d0607e85f5e44774284628227ea994d4, type: 3}
       propertyPath: m_RootOrder
@@ -593,6 +634,18 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 39023575}
     m_Modifications:
+    - target: {fileID: -4434437990185676618, guid: 89e09b3948d61ed449826f468c84e909, type: 3}
+      propertyPath: m_IsTrigger
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1108470480895323624, guid: 89e09b3948d61ed449826f468c84e909, type: 3}
+      propertyPath: m_Name
+      value: pochag
+      objectReference: {fileID: 0}
+    - target: {fileID: 1108470480895323624, guid: 89e09b3948d61ed449826f468c84e909, type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
     - target: {fileID: 1726430968732564686, guid: 89e09b3948d61ed449826f468c84e909, type: 3}
       propertyPath: m_RootOrder
       value: 0
@@ -641,7 +694,17 @@ PrefabInstance:
       propertyPath: m_Name
       value: dager
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    - target: {fileID: 3107797298255059210, guid: 89e09b3948d61ed449826f468c84e909, type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3107797298255059210, guid: 89e09b3948d61ed449826f468c84e909, type: 3}
+      propertyPath: m_TagString
+      value: SubAttack
+      objectReference: {fileID: 0}
+    m_RemovedComponents:
+    - {fileID: 0}
+    - {fileID: 0}
   m_SourcePrefab: {fileID: 100100000, guid: 89e09b3948d61ed449826f468c84e909, type: 3}
 --- !u!1 &1684170819
 GameObject:
@@ -777,6 +840,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   basicAttackPrefab: {fileID: 3107797298255059210, guid: 89e09b3948d61ed449826f468c84e909, type: 3}
+  speed: 20
 --- !u!1 &1859014899 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 2314681825701584322, guid: 129f1b5fe10ef424da0ff2abc5981855, type: 3}
