+++ My project/ProjectSettings/TagManager.asset
@@ -8,6 +8,7 @@ TagManager:
   - AttackRange
   - skill
   - Monster Group
+  - SubAttack
   layers:
   - Default
   - TransparentFX
